fix: separate factory method to another file
